Avoid the false debugging message
Any time the editor was used, the following debugging message was
displayed:

```
Too many params passed to js_call_amd("editor_marklar/editor", "init")
```

The huge options must get into the browser though as it is part of the
legacy filepicker infrastructure. This patch introduces a nasty hack
that allows us to bypass the init params size check as the file picker
options are not passed directly but read from a global variable.

See MDL-53423 for details.